% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.report.R
\name{anova_report}
\alias{anova_report}
\title{Compare Data with One-Way ANOVA}
\usage{
anova_report(x, units = "")
}
\arguments{
\item{x}{data file, extracted from LIMS.}

\item{units}{units of measure used in data.}
}
\description{
This function produces a word document comparing two or more datasets by ANOVA.
}
\keyword{ANOVA}

