% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combiner.R
\name{combo}
\alias{combo}
\title{A routine for creating repeatability and reproducibility combination pairs from multiple data batches.
The data is in similar format to ANOVA data.}
\usage{
combo(x)
}
\arguments{
\item{x}{a csv file of multiple batch data, columns headed but rows unnamed.  No row or column limits.}
}
\description{
A routine for creating repeatability and reproducibility combination pairs from multiple data batches.
The data is in similar format to ANOVA data.
}
\keyword{MU}
\keyword{combinations,}
