% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupes.R
\name{dupes}
\alias{dupes}
\title{This function reduces a file to the duplicated values only, using the nominated column or columns.}
\usage{
dupes(x, y)
}
\arguments{
\item{x}{Data frame name.}

\item{y}{Nominated column(s) - use c("A","B") format.}
}
\description{
This function reduces a file to the duplicated values only, using the nominated column or columns.
}
\keyword{duplicates}

