% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horwitz.R
\name{horwitz}
\alias{horwitz}
\title{Horwitz Value}
\usage{
horwitz(x, n)
}
\arguments{
\item{x}{- a value to be assessed.}

\item{n}{- conversion to g/g.  n = 2 for g/100g, 6 for mg/kg.}
}
\description{
This function calculates a Horwitz standard deviation for a given data value.
}
\keyword{horwitz}
