% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outliers.R
\name{outliers}
\alias{outliers}
\title{A routine for identifying outliers in a vector of data based on IQR robust statistics.}
\usage{
outliers(x, b = FALSE)
}
\arguments{
\item{x}{a data set.}

\item{b}{output boxplot of the four sweeps to remove outliers.  Default is FALSE.}
}
\description{
A routine for identifying outliers in a vector of data based on IQR robust statistics.
}
\keyword{MU}
\keyword{outliers,}

