% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100g.adjust.R
\name{per100_data}
\alias{per100_data}
\title{This function creates uniform data units in an MU data file.}
\usage{
per100_data(x)
}
\arguments{
\item{x}{MU data file, extracted from LIMS and cleaned.}
}
\description{
This function creates uniform data units in an MU data file.
}
\keyword{cleaning}
\keyword{data}

