% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1kg.adjust.R
\name{perkg_data}
\alias{perkg_data}
\title{This function creates uniform mg/kg data units in an MU data file.}
\usage{
perkg_data(x)
}
\arguments{
\item{x}{MU data file, extracted from LIMS and cleaned.}
}
\description{
This function creates uniform mg/kg data units in an MU data file.
}
\keyword{cleaning}
\keyword{data}

