% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MU.main.calc.R
\name{process.MU}
\alias{process.MU}
\title{This function does the core MU calculations.}
\usage{
process.MU(x, hv)
}
\arguments{
\item{x}{MU data file, extracted from LIMS and cleaned.}

\item{hv}{Horwitz value (g/100g = 2, mg/kg = 6, no units = 0)}
}
\description{
This function does the core MU calculations.
}
\keyword{MU}
\keyword{calculations}
