% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.limits.R
\name{srm.limits}
\alias{srm.limits}
\title{This function takes the srm file stripped from the MU calculation data and determines associated quality parameters.}
\usage{
srm.limits(x, dp = 3)
}
\arguments{
\item{x}{Data file (csv) exported from MU calculation routine.}

\item{dp}{Decimal places in the summary table.  Default = 3.}
}
\description{
This function takes the srm file stripped from the MU calculation data and determines associated quality parameters.
}
\keyword{MU}
\keyword{chart,}
\keyword{control}
\keyword{srm,}

