% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm.report.R
\name{srm_report}
\alias{srm_report}
\title{Analyse SRM Data}
\usage{
srm_report(x, max.pts = 100, points = 20)
}
\arguments{
\item{x}{data file, extracted from LIMS.}

\item{max.pts}{number of points used in control chart.}

\item{points}{number of points at the start of the control chart used to calculate control lines.}
}
\description{
This function produces a word document assessing an SRM performance.
}
\keyword{SRM,}
\keyword{charts}
\keyword{control,}

