% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MU.clean.R
\name{summary_data}
\alias{summary_data}
\title{This function summarises data in a extracted csv file to display numbers of reported names and units.}
\usage{
summary_data(x)
}
\arguments{
\item{x}{MU data file, extracted from LIMS.}
}
\description{
This function summarises data in a extracted csv file to display numbers of reported names and units.
}
\keyword{cleaning}
\keyword{data}

