% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCAR.extract.R
\name{TCAR}
\alias{TCAR}
\title{A routine for investigating Total CHO anomolies.}
\usage{
TCAR(x)
}
\arguments{
\item{x}{a dataframe, extracted from LIMS}
}
\description{
A routine for investigating Total CHO anomolies.
}
\keyword{Carbohydrates}
\keyword{TCAR,}
\keyword{Total}

