% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOST.R
\name{TOST}
\alias{TOST}
\title{TOST Equivalence Test}
\usage{
TOST(A, B, E, b = FALSE)
}
\arguments{
\item{A}{- First data set.}

\item{B}{- Second data set.}

\item{E}{- acceptable difference between the sets.}

\item{b}{- option boxplot output, default = FALSE}
}
\description{
This function calculates whether two sets of data are practically equivalent.
}
\keyword{TOST,}
\keyword{equivalence,}

