% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_list.R
\name{weekly_list}
\alias{weekly_list}
\title{This function creates a summary file based on the exported data from Google Kanbanchi.}
\usage{
weekly_list(x)
}
\arguments{
\item{x}{A variable equal to the Kanbanchi export file "LK_Projects_full.csv"}
}
\description{
This function creates a summary file based on the exported data from Google Kanbanchi.
}
\keyword{kanbanchi,}
\keyword{projects}
